feat: add testimonials section for enhanced client feedback presentation
